Rename environment variables & bump version.
It turns out, `UWSGI_VERSION` as a defined environment makes uwsgi ...
not work. It makes it just print out the version (ala a --version flag).
So after many hours of hair-pulling, let's rename all our env variables
that we use to define which versions of our libraries/dependencies we
want to use with a `DEP_` prefix. I've gone for `DEP` rather than `DM`
as we use the latter for environment variables that feed into our apps,
whereas these are just building the container, so it feels good to try
to avoid namespace clashes.

Also bump to version 3.1.1 so we can actually use the updates.