Tag images with build date
This commit changes the Makefile so that images have an additional tag
that includes the build date. This allows a simple history to be kept
for when the base image changes underneath the version number.

It also means that we can get the latest and greatest base image without
changing the version numbers of our images.