supervisord: attempt more graceful shutdown procedure through better use of signals
we will never make this perfect as it's possible for an instance to have
a significant number of queued connections and only a 10s grace period
given by bosh in which to clear that backlog. so largely make daemons
ignore the stop signal and let them get on with clearing that backlog.