Upgrade to Python 3.7
Python 3.6 goes end of life at the end of the year. There are also some new features in 3.7 that we'd like, such as data classes.

We run all our repo CI in all Pythons between 3.6 and 3.9, so I'm not expecting any incompatibilities.

Create base images using Python 3.7 that we can use for testing. This commit shouldn't be merged to master as-is. Test first in Preview, and then update the VERSION file before releasing further.